// binders-devops-service-v1/app/src/dc/docker-compose-spec.d.ts
export type Include =
    | string
    | {
          path?: StringOrList;
          env_file?: StringOrList;
          project_directory?: string;
      };
export type StringOrList = string | ListOfStrings;
// ...
export type ListOrDict =
    | {
          /**
           * This interface was referenced by `undefined`'s JSON-Schema definition
           * via the `patternProperty` ".+".
           */
          [k: string]: string | number | boolean | null;
      }
    | string[];
// ...
export type Network = {
    name?: string;
    driver?: string;
    driver_opts?: {
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^.+$".
         */
        [k: string]: string | number;
    };
    ipam?: {
        driver?: string;
        config?: {
            subnet?: string;
            ip_range?: string;
            gateway?: string;
            aux_addresses?: {
                /**
                 * This interface was referenced by `undefined`'s JSON-Schema definition
                 * via the `patternProperty` "^.+$".
                 */
                [k: string]: string;
            };
            /**
             * This interface was referenced by `undefined`'s JSON-Schema definition
             * via the `patternProperty` "^x-".
             */
            [k: string]: unknown;
        }[];
        options?: {
            /**
             * This interface was referenced by `undefined`'s JSON-Schema definition
             * via the `patternProperty` "^.+$".
             */
            [k: string]: string;
        };
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^x-".
         */
        [k: string]: unknown;
    };
    external?:
        | boolean
        | string
        | {
              /**
               * @deprecated
               */
              name?: string;
              /**
               * This interface was referenced by `undefined`'s JSON-Schema definition
               * via the `patternProperty` "^x-".
               */
              [k: string]: unknown;
          };
    internal?: boolean | string;
    enable_ipv4?: boolean | string;
    enable_ipv6?: boolean | string;
    attachable?: boolean | string;
    labels?: ListOrDict;
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     *
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     */
    [k: string]: unknown;
} & Network1;
export type Network1 = {
    name?: string;
    driver?: string;
    driver_opts?: {
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^.+$".
         */
        [k: string]: string | number;
    };
    ipam?: {
        driver?: string;
        config?: {
            subnet?: string;
            ip_range?: string;
            gateway?: string;
            aux_addresses?: {
                /**
                 * This interface was referenced by `undefined`'s JSON-Schema definition
                 * via the `patternProperty` "^.+$".
                 */
                [k: string]: string;
            };
            /**
             * This interface was referenced by `undefined`'s JSON-Schema definition
             * via the `patternProperty` "^x-".
             */
            [k: string]: unknown;
        }[];
        options?: {
            /**
             * This interface was referenced by `undefined`'s JSON-Schema definition
             * via the `patternProperty` "^.+$".
             */
            [k: string]: string;
        };
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^x-".
         */
        [k: string]: unknown;
    };
    external?:
        | boolean
        | string
        | {
              /**
               * @deprecated
               */
              name?: string;
              /**
               * This interface was referenced by `undefined`'s JSON-Schema definition
               * via the `patternProperty` "^x-".
               */
              [k: string]: unknown;
          };
    internal?: boolean | string;
    enable_ipv4?: boolean | string;
    enable_ipv6?: boolean | string;
    attachable?: boolean | string;
    labels?: ListOrDict;
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     *
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     */
    [k: string]: unknown;
} | null;
/**
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "^[a-zA-Z0-9._-]+$".
 */
export type Volume = {
    name?: string;
    driver?: string;
    driver_opts?: {
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^.+$".
         */
        [k: string]: string | number;
    };
    external?:
        | boolean
        | string
        | {
              /**
               * @deprecated
               */
              name?: string;
              /**
               * This interface was referenced by `undefined`'s JSON-Schema definition
               * via the `patternProperty` "^x-".
               */
              [k: string]: unknown;
          };
    labels?: ListOrDict;
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     *
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     */
    [k: string]: unknown;
} & Volume1;
export type Volume1 = {
    name?: string;
    driver?: string;
    driver_opts?: {
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^.+$".
         */
        [k: string]: string | number;
    };
    external?:
        | boolean
        | string
        | {
              /**
               * @deprecated
               */
              name?: string;
              /**
               * This interface was referenced by `undefined`'s JSON-Schema definition
               * via the `patternProperty` "^x-".
               */
              [k: string]: unknown;
          };
    labels?: ListOrDict;
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     *
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     */
    [k: string]: unknown;
} | null;

/**
 * The Compose file is a YAML file defining a multi-containers based application.
 */
export interface ComposeSpecification {
    /**
     * declared for backward compatibility, ignored.
     */
    version?: string;
    /**
     * define the Compose project name, until user defines one explicitly.
     */
    name?: string;
    /**
     * compose sub-projects to be included.
     */
    include?: Include[];
    services?: {
        [k: string]: Service;
    };
    networks?: {
        [k: string]: Network;
    };
    volumes?: {
        [k: string]: Volume;
    };
    secrets?: {
        [k: string]: Secret;
    };
    configs?: {
        [k: string]: Config;
    };
    /**
     * This interface was referenced by `ComposeSpecification`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     */
    [k: string]: unknown;
}
// ...
export interface Secret {
    name?: string;
    environment?: string;
    file?: string;
    external?:
        | boolean
        | string
        | {
              name?: string;
              [k: string]: unknown;
          };
    labels?: ListOrDict;
    driver?: string;
    driver_opts?: {
        /**
         * This interface was referenced by `undefined`'s JSON-Schema definition
         * via the `patternProperty` "^.+$".
         */
        [k: string]: string | number;
    };
    template_driver?: string;
    /**
     * This interface was referenced by `Secret`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     */
    [k: string]: unknown;
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "^[a-zA-Z0-9._-]+$".
 */
export interface Config {
    name?: string;
    content?: string;
    environment?: string;
    file?: string;
    external?:
        | boolean
        | string
        | {
              /**
               * @deprecated
               */
              name?: string;
              [k: string]: unknown;
          };
    labels?: ListOrDict;
    template_driver?: string;
    /**
     * This interface was referenced by `Config`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     */
    [k: string]: unknown;
}